requirements.txt
# CAMEL-AI框架核心依赖
camel-ai[all]>=0.1.5

# 向量数据库
qdrant-client>=1.7.0

# 向量化模型和深度学习
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0

# Web界面
streamlit>=1.28.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 文本处理
nltk>=3.8.0
jieba>=0.42.1

# HTTP请求和API
requests>=2.31.0
openai>=1.3.0

# 工具库
tqdm>=4.66.0
python-dotenv>=1.0.0

# 可选：GPU加速（如果有NVIDIA GPU）
# torch-audio>=2.0.0
# torch-vision>=0.15.0

# 可选：更多文档格式支持
# pypdf>=3.17.0
# python-docx>=0.8.11
# beautifulsoup4>=4.12.0

# 加载环境变量
python-dotenv==1.0.0

# 开发和测试工具
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0